Refactor: Move special metrics (percentage, conversion rate) to own module
