+++ lib/plausible/stats/base.ex
@@ -139,84 +139,4 @@ defmodule Plausible.Stats.Base do
 
     "^#{escaped}$"
   end
-
-  defp total_visitors(site, query) do
-    base_event_query(site, query)
-    |> select([e],
-      total_visitors: fragment("toUInt64(round(uniq(?) * any(_sample_factor)))", e.user_id)
-    )
-  end
-
-  # `total_visitors_subquery` returns a subquery which selects `total_visitors` -
-  # the number used as the denominator in the calculation of `conversion_rate` and
-  # `percentage` metrics.
-
-  # Usually, when calculating the totals, a new query is passed into this function,
-  # where certain filters (e.g. goal, props) are removed. That might make the query
-  # able to include imported data. However, we always want to include imported data
-  # only if it's included in the base query - otherwise the total will be based on
-  # a different data set, making the metric inaccurate. This is why we're using an
-  # explicit `include_imported` argument here.
-  def total_visitors_subquery(site, query, include_imported)
-
-  def total_visitors_subquery(site, query, true = _include_imported) do
-    wrap_expression([], %{
-      total_visitors:
-        subquery(total_visitors(site, query)) +
-          subquery(Plausible.Stats.Imported.total_imported_visitors(site, query))
-    })
-  end
-
-  def total_visitors_subquery(site, query, false = _include_imported) do
-    wrap_expression([], %{
-      total_visitors: subquery(total_visitors(site, query))
-    })
-  end
-
-  def add_percentage_metric(q, site, query, metrics) do
-    if :percentage in metrics do
-      total_query = Query.set_dimensions(query, [])
-
-      q
-      |> select_merge_as([], total_visitors_subquery(site, total_query, query.include_imported))
-      |> select_merge_as([], %{
-        percentage:
-          fragment(
-            "if(? > 0, round(? / ? * 100, 1), null)",
-            selected_as(:total_visitors),
-            selected_as(:visitors),
-            selected_as(:total_visitors)
-          )
-      })
-    else
-      q
-    end
-  end
-
-  # Adds conversion_rate metric to query, calculated as
-  # X / Y where Y is the same breakdown value without goal or props
-  # filters.
-  def maybe_add_conversion_rate(q, site, query, metrics) do
-    if :conversion_rate in metrics do
-      total_query =
-        query
-        |> Query.remove_filters(["event:goal", "event:props"])
-        |> Query.set_dimensions([])
-
-      # :TRICKY: Subquery is used due to event:goal breakdown above doing an UNION ALL
-      subquery(q)
-      |> select_merge_as([], total_visitors_subquery(site, total_query, query.include_imported))
-      |> select_merge_as([e], %{
-        conversion_rate:
-          fragment(
-            "if(? > 0, round(? / ? * 100, 1), 0)",
-            selected_as(:total_visitors),
-            e.visitors,
-            selected_as(:total_visitors)
-          )
-      })
-    else
-      q
-    end
-  end
 end
