+++ lib/plausible/stats/sql/fragments.ex
@@ -203,4 +203,19 @@ defmodule Plausible.Stats.SQL.Fragments do
   end
 
   defp update_literal_map_values(ast, _), do: ast
+
+  @doc """
+  Macro that helps join two Ecto queries by selecting fields from either one
+  """
+  defmacro select_join_fields(q, query, list, table_name) do
+    quote do
+      Enum.reduce(unquote(list), unquote(q), fn metric_or_dimension, q ->
+        key = shortname(unquote(query), metric_or_dimension)
+
+        select_merge_as(q, [e, s], %{
+          key => field(unquote(table_name), ^key)
+        })
+      end)
+    end
+  end
 end
