+++ lib/plausible/stats/sql/query_builder.ex
@@ -43,9 +43,7 @@ defmodule Plausible.Stats.SQL.QueryBuilder do
     |> join_sessions_if_needed(site, events_query)
     |> build_group_by(events_query)
     |> merge_imported(site, events_query, events_query.metrics)
-    |> maybe_add_global_conversion_rate(site, events_query)
-    |> maybe_add_group_conversion_rate(site, events_query)
-    |> Base.add_percentage_metric(site, events_query, events_query.metrics)
+    |> SQL.SpecialMetrics.add(site, events_query)
   end
 
   defp join_sessions_if_needed(q, site, query) do
@@ -86,9 +84,7 @@ defmodule Plausible.Stats.SQL.QueryBuilder do
     |> join_events_if_needed(site, sessions_query)
     |> build_group_by(sessions_query)
     |> merge_imported(site, sessions_query, sessions_query.metrics)
-    |> maybe_add_global_conversion_rate(site, sessions_query)
-    |> maybe_add_group_conversion_rate(site, sessions_query)
-    |> Base.add_percentage_metric(site, sessions_query, sessions_query.metrics)
+    |> SQL.SpecialMetrics.add(site, sessions_query)
   end
 
   def join_events_if_needed(q, site, query) do
@@ -153,87 +149,6 @@ defmodule Plausible.Stats.SQL.QueryBuilder do
         selected_as(^shortname(query, metric_or_dimension))
       }
     )
-  end
-
-  defmacrop select_join_fields(q, query, list, table_name) do
-    quote do
-      Enum.reduce(unquote(list), unquote(q), fn metric_or_dimension, q ->
-        key = shortname(unquote(query), metric_or_dimension)
-
-        select_merge_as(q, [e, s], %{
-          key => field(unquote(table_name), ^key)
-        })
-      end)
-    end
-  end
-
-  # Adds conversion_rate metric to query, calculated as
-  # X / Y where Y is the same breakdown value without goal or props
-  # filters.
-  def maybe_add_global_conversion_rate(q, site, query) do
-    if :conversion_rate in query.metrics do
-      total_query =
-        query
-        |> Query.remove_filters(["event:goal", "event:props"])
-        |> Query.set_dimensions([])
-
-      q
-      |> select_merge_as(
-        [],
-        Base.total_visitors_subquery(site, total_query, query.include_imported)
-      )
-      |> select_merge_as([e], %{
-        conversion_rate:
-          fragment(
-            "if(? > 0, round(? / ? * 100, 1), 0)",
-            selected_as(:total_visitors),
-            selected_as(:visitors),
-            selected_as(:total_visitors)
-          )
-      })
-    else
-      q
-    end
-  end
-
-  # This function injects a group_conversion_rate metric into
-  # a dimensional query. It is calculated as X / Y, where:
-  #
-  #   * X is the number of conversions for a set of dimensions
-  #     result (conversion = number of visitors who
-  #     completed the filtered goal with the filtered
-  #     custom properties).
-  #
-  #  * Y is the number of all visitors for this set of dimensions
-  #    result without the `event:goal` and `event:props:*`
-  #    filters.
-  def maybe_add_group_conversion_rate(q, site, query) do
-    if :group_conversion_rate in query.metrics do
-      group_totals_query =
-        query
-        |> Query.remove_filters(["event:goal", "event:props"])
-        |> Query.set_metrics([:visitors])
-        |> Query.set_order_by([])
-
-      from(e in subquery(q),
-        left_join: c in subquery(build(group_totals_query, site)),
-        on: ^build_group_by_join(query)
-      )
-      |> select_merge_as([e, c], %{
-        total_visitors: c.visitors,
-        group_conversion_rate:
-          fragment(
-            "if(? > 0, round(? / ? * 100, 1), 0)",
-            c.visitors,
-            e.visitors,
-            c.visitors
-          )
-      })
-      |> select_join_fields(query, query.dimensions, e)
-      |> select_join_fields(query, List.delete(query.metrics, :group_conversion_rate), e)
-    else
-      q
-    end
   end
 
   defp join_query_results({nil, _}, {nil, _}), do: nil
